add example/index.html
