+++ example/001/bundles/inject.js
