+++ example/001/index.html
@@ -9,12 +9,13 @@
             };
         </script>
         <script src="../bower_components/injectorjs/dist/injector.js"></script>
-        <script>
-          FAI.INJECTOR.registerEvents(["DOMContentLoaded"]);
-        </script>
     </head>
     <body>
         <h1>Injectorjs</h1>
-        <div id="output" data-bundle="inject"></div>
+        <h2>Inject js with javascript</h2>
+        <div id="output"></div>
+        <script>
+            FAI.INJECTOR.loadFeature('inject');
+        </script>
     </body>
 </html>
