+++ example/002/bundles/inject.js
