+++ example/003/index.html
@@ -15,6 +15,7 @@
     </head>
     <body>
         <h1>Injectorjs</h1>
+        <h2>Inject webpack bundle</h2>
         <div id="output" data-bundle="inject"></div>
     </body>
 </html>
