+++ example/003/package.json
