+++ example/003/src/inject.js
