+++ example/004/bundles/hellos.js
