+++ example/004/src/inject.js
