+++ example/004/src/inyectar.js
