+++ example/005/index.html
@@ -15,6 +15,7 @@
     </head>
     <body data-bundle='hellos'>
         <h1>Injectorjs</h1>
+        <h2>Inject webpack bundles (css, html, and js)</h2>
         <div id="output"></div>
     </body>
 </html>
