+++ example/005/src/js/renderer.js
