+++ example/README.md
@@ -1,1 +1,23 @@
+# Examples
+
 To run these examples start a server (any port) with this as the root directory.
+
+## Example 1
+
+Inject js with javascript
+
+## Example 2
+
+Inject js with data tag
+
+## Example 3
+
+Inject webpack bundle
+
+## Example 4
+
+Inject webpack bundles
+
+## Example 5
+
+Inject webpack bundles that include css, html, and js
