+++ example/index.html
@@ -1,0 +1,15 @@
+<html>
+    <head>
+        <title>Injectorjs</title>
+    </head>
+    <body data-bundle='hellos'>
+        <h1>Injectorjs</h1>
+        <ol>
+            <li><a href="001">Inject js with javascript</a></li>
+            <li><a href="002">Inject js with data tag</a></li>
+            <li><a href="003">Inject webpack bundle</a></li>
+            <li><a href="004">Inject webpack bundles</a></li>
+            <li><a href="005">Inject webpack bundles css html js</a></li>
+        </ol>
+    </body>
+</html>
